requirements.txt
streamlit
openai-whisper
google-generativeai
dnspython
pymongo
streamlit-authenticator
pymongo 
bcrypt
certifi
extra-streamlit-components